// web/lib/api/generated/model/dispositionEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `RETURN_TO_STOCK` - Return to stock
 * * `DAMAGED` - Received damaged
 * * `NOT_RETURNED` - Not physically returned
 */
export type DispositionEnum = typeof DispositionEnum[keyof typeof DispositionEnum];


export const DispositionEnum = {
  RETURN_TO_STOCK: 'RETURN_TO_STOCK',
  DAMAGED: 'DAMAGED',
  NOT_RETURNED: 'NOT_RETURNED',
} as const;