// web/lib/api/generated/model/creditNoteCreateKindEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `RETURN` - RETURN
 * * `PRICE_ADJUSTMENT` - PRICE_ADJUSTMENT
 */
export type CreditNoteCreateKindEnum = typeof CreditNoteCreateKindEnum[keyof typeof CreditNoteCreateKindEnum];


export const CreditNoteCreateKindEnum = {
  RETURN: 'RETURN',
  PRICE_ADJUSTMENT: 'PRICE_ADJUSTMENT',
} as const;