// web/lib/api/generated/model/creditNotesListKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

export type CreditNotesListKind = typeof CreditNotesListKind[keyof typeof CreditNotesListKind];


export const CreditNotesListKind = {
  CANCELLATION: 'CANCELLATION',
  PRICE_ADJUSTMENT: 'PRICE_ADJUSTMENT',
  RETURN: 'RETURN',
  SHORT_SUPPLY: 'SHORT_SUPPLY',
} as const;